LPS-157265 Do for 74
